refactor(test): use globalroutes file instead index file
