+++ src/modules/home/home.test.ts
@@ -1,4 +1,4 @@
-import app from "../../server";
+import app from "../../server/_globalRoutes";
 
 describe("Home routes tester", () => {
